mononoke/repo_client: support ChangegroupV3 push
Summary:
We have most of the support for ChangegroupV3 in place already (we've been
using it commit cloud for a while, and we have tests for pushrebase). However,
we didn't really support push because we are looking for the wrong field in the
wrong set of headers (and we don't advertise V3 in capabilities).

This fixes that. It does make us a little more lenient by allowing the version
in either advisory and mandatory params (mainly because our commit cloud sync
and Mercurial itself look like they're doing something different there).

Note: I removed the constant names since the field name is really a thing that
only makes sense for each individual part type, so it seemed like it made more
sense to have the 2 right next to each other.

Reviewed By: ikostia

Differential Revision: D17982127

fbshipit-source-id: 32f822f7aa5fcb7ca5eded5d0cd5fff4af36e606