types: rename Node to HgId
Summary:
`Node` is a name that is overloaded. It means that in our conversations
we always have to constantly define which Node we are talking about.
Is it the type in Mercurial or the generic term in graphs/trees.

It is just an identifier so we rename it such.

Note that this doesn't remove the Node term yet. I will take care of the
longtail of uses incrementally.

Things that are not renamed:
* linknode
* nodeinfo
* WireHistoryEntry
* cdatapack

Reviewed By: quark-zju

Differential Revision: D18010149

fbshipit-source-id: 83b3911e231e4544848391cd3deb6e44ec2b809d