types: add serde rename and alias for Key
Summary:
"node" is used in the wire protocol.
While things are deploying the server needs to receive `node`. For
that we use `serde(rename = "node")`.
Looking forward we will need the server to know how to interpret both
`node` and `hgid` while clients are deploying without
`serde(rename = "node")`. For that we use `serde(alias = "node")`.

Reviewed By: kulshrax

Differential Revision: D18041109

fbshipit-source-id: 06b1568ad010526a53d2ccfd46950211d303250d