Convert repo source: Fix getchanges() and some cleanup
Summary: We've had a known problem with importing merge commits in the currently landed version of the repo converter source. The problem is that the specification of the source interface is misleading (or perhaps just wrong). We've updated the function to generate changes against all of the parents, not just the first one, and that seems to have fixed our ability to show the head commit of branches.

Reviewed By: tchebb

Differential Revision: D17964281

fbshipit-source-id: 90ae5df7dabaf133612c44d2a41b37200bcdb15f