generation of diff header
Summary:
Added function generates a basic diff header. I'll follow up this diff with additions of:
 * git-specific headers (`diff --git`): file modes, copy information
 * no-prefix mode (without `a/` and `b/`)

High-level context: I'm working on adding support for returning diff from the new Source Control Service. We want the generated diff to be a drop-in replacement for the one that we get from hg in SCMQuery so I'm closely following the behaviour of Mercurial's `mdiff.py` so there's no surprises. I'm planning to implement all of the diff features that we'll use in the Service while making the implementation flexible enough that it can be easily extended in the future and dropped into Rust rewrite of Mercurial.

Reviewed By: markbt

Differential Revision: D17764609

fbshipit-source-id: 43cde9ddef86d01fe926adedb3f2ac19b38ffbec