generation of headerless diff
Summary:
This code, loosely inspired Mercurial's mdiff.py generates the actual "diff"
contents (everything that we can generate just based on the file contents only,
without looking at file metadata such as: filenames, copy information, file
modes etc. Those will follow in separate diffs.

High-level context: I'm working on adding support for returning diff from the new Source Control Service. We want the generated diff to be a drop-in replacement for the one that we get from hg in SCMQuery. I'm planning to implement all of the diff features that we'll use in the Service while making the implementation flexible enough that it can be easily extended in the future and dropped into Rust rewrite of Mercurial.

Reviewed By: krallin

Differential Revision: D17764608

fbshipit-source-id: 4fa3a1afd0547fc512c5e51ed673b1ea007b3245