cargo_from_buck: add support for generating Cargo.toml from thrift_library
Summary:
In order to get thrift to work I've used a feature of cargo called "build script".
The idea is that in addition to "Cargo.toml" two files are generated for thrift_library:
  - build.rs - it is being called during build time, it calls "buck build" that generates the thrift code and copies it over to Cargo's target folder
  - lib.rs - it includes the content of the generated file, as if the content of the generated file was the content of this lib.rs
All these together make it possible to leverage buck to generate the rust thrift code and build it with everything else.

Reviewed By: farnz

Differential Revision: D17950013

fbshipit-source-id: 2a6bbc537ed7fb5a49461f12032d3be6748e0fa4