util: add expand_path function
Summary: Add an `expand_path` function to the `util` crate, similar to the `util.expandpath` Python function. This function is similar to one from the `configparser` crate (which it now replaces). A key difference is that this function is designed to be similar to the Python function in terms of expanding home directory references in environment variables.

Reviewed By: xavierd

Differential Revision: D18050348

fbshipit-source-id: 752d85864eaa4c296ca670089f5de2dc860a5d6a