mononoke: don't use lazy_static in signal context
Summary:
I'm pretty sure lazy_static's mechanism isn't signal-safe. In this
case it probably doesn't matter since the lazy_static is materialized before
the signal handler is registered, so there's no chance of doing the
materialization in signal context, but its best not to risk it. There's no need
for a lazy_static for AtomicBool, so why take the risk?

Reviewed By: StanislavGlebik

Differential Revision: D18066064

fbshipit-source-id: e12e429080f83628829e28a0d2bb0dd7a44542f9