mononoke: backsyncer with commit rewriting
Summary:
Use functionality added in D18121305 to implement backsyncer that knows how to
rewrite commits (previously it was just copying commits from one repo to
another).

Note that tests were rewritten from scratch, and one of the tests was removed
without replacement - the one that tests backsyncer on diamond merges. It won't
work right now because there are known innefficiencies in
`get_commit_sync_outcome()`, which will be addressed later.

Reviewed By: farnz

Differential Revision: D18134820

fbshipit-source-id: f1ee39842f7e89d057d5be58667dab739df97b95