cmdlib: make MononokeApp interface more fluent
Summary:
The `MononokeApp` interface is almost like a builder.  Make it a proper builder
with a fluent API.  This will allow us to add new options in the future without
changing all the places it is called.

As an example of why this is useful, we also clean up the options for selecting
a repo.  The default behaviour is to add `--repo-id` and `--repo-name` options
that are optional.  This is not usually what programs want, so they can now
specify `.with_repo_required()` to make one of these options mandatory, or
`.with_all_repos()` to remove the options entirely for commands that operate
on all repos.

The LFS server is updates to specify `.with_all_repos()`.  We will update the
remaining existing commands separately as needed.

Reviewed By: krallin

Differential Revision: D18009138

fbshipit-source-id: 5ce46ee2418c92c3697138cccb2b3e2807e46faa