configparser: implement `FromConfigValue` for `PathBuf`
Summary:
On macOS / CPE managed Linux we set `remotefilelog.cachepath` to `~/.hgcache`. However, `contentstore` does not know how to expand tilde sign, and this will result as a file not found error.

This diff implements `FromConfigValue` for `PathBuf` with `expand_path`, so we can use `get_or::<PathBuf>(...)` when we are expecting a path from the config.

Reviewed By: xavierd

Differential Revision: D18128810

fbshipit-source-id: b489a91116b6d0424c5e4675c8d0ae9775242cbc