cpython-ext: make it compile with python3
Summary:
The PyBuffer_Check function no longer exist in Python3, and I'm not sure if an
equivalent will be needed in Python3. For now, let's simply not call it, once
we get Mercurial running in Python3, the code will be overly pessimistic and
thus may trigger the panic code more often than in Python2, we'll see what to
do then.

Reviewed By: quark-zju

Differential Revision: D18183657

fbshipit-source-id: 06c9c82402be28c8121c063d4a4ec3e3b397b4b5