build: make it easier to switch between python2/3
Summary:
The python version is now a variable in targets.bzl that control which python
Rust crate are being used. This will make it easier to try to get Mercurial
compiled against python3 in the future.

For now, in the python3 mode, Mercurial doesn't compile, due mostly to the
following:
 - cstore, will be removed once the Rust manifest code is everywhere,
 - cfastmanifest, no longer necessary as hybridmanifest are no longer used
 - couple of Rust crate that use Python2 only constructs

Reviewed By: quark-zju

Differential Revision: D18120177

fbshipit-source-id: d63188fa01081f8f6a5bfff98001e1c584c545d7