checkmessagehook: allow non-printable characters
Summary:
While we do want to prevent users from adding non-printable characters to the
commit message, importing commits from other repos may conflict with this rule.
In this case, disabling the hook altogether is not an option as non UTF8 in the
commit message will prevent pushing to the server.

Reviewed By: tchebb

Differential Revision: D18192039

fbshipit-source-id: 8a9aab077cf20826d4885841df5e9b7dd033b598