process renames in file history
Summary:
If unode doesn't have parents, it means that either the file was just created or it was copied/renamed.

For each unode's changeset that doesn't have parents I check whether files was renamed. If so, I proceed bounded traversal to the next stage with new path and changeset id from `copy_from` information.

Reviewed By: StanislavGlebik

Differential Revision: D18138431

fbshipit-source-id: 9e499fa9003a1cfb9ab7400bf6a609f174f7f676