file_history: keep potential source nodes for the commit graph traversal
Summary:
I collect first nodes for all unode batches bounded traversal fetches with the stages (depth) where they were fetched.
These nodes are potential sources in the commit graph, where "source" - node that is not a parent for any commit in a graph.

This allows me to not fetch again the very first unode after the bounded traversal, but just get it from the collected potential changesets.

Reviewed By: StanislavGlebik

Differential Revision: D18118155

fbshipit-source-id: 41a835c71c77b65951af211b046fd13f91fdcf72