templater: add hyperlink function to print hyperlink
Summary:
(I'm not sure if `templater.py` is a good place for this function. Let me know if there is a better place for it)

See https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda for background on hyperlinks in terminal.

This diff adds a template function that allows us to render hyper links in terminal (aka clickable URL without printing the URL).

Reviewed By: quark-zju

Differential Revision: D18125244

fbshipit-source-id: eca1d05070b0d49545eecad8f8ef8add13140d08