mononoke: fix pushrebase when pushing branch merge
Summary:
At the moment all "branch merge" pushrebases were able to corrupt a repo
because of the way bonsai commits work. In particular, if a file A was modified
between pushrebase root and "onto" commit (e.g. master) and this file A existed
before root commit, then the merged bonsai commit must have a file change for
file A to resolve a conflict between p1 and p2 of a merge commit.

   C <- master (pushrebase onto, E will be rebased here)
   :
   | E <- rev being pushed, a merge
   |/:
   B | <- B is the pushrebase "root" (???), since E is being rebased off B.
   : D <- the server does not have D, 'E % C', aka. D and E are being pushed.
   A |
   :\|
   | C <- the server has A, so it has C
   ~

This diff fixes it by explicitly adding changed files to the pushrebased bonsai
changeset.

Note that new changed files are added both for "branch merge" and for "repo
merge" (i.e. when ancestors of p2 do not intersect with ancestors p1), and this
is not necessary for the latter case, but it should still be correct.

Reviewed By: krallin

Differential Revision: D18065013

fbshipit-source-id: 5c73b43ff1e75040cc9b2c39f8a87f075f77e345