hggit: correct import error handling for test-hggit-url-parsing
Summary:
Formatting the code with Black changed the import order for this test, moving
the test for dulwich's existence to *after* the import that used it.

This means the test will fail on platforms where dulwich is not available.

Move the import that actually uses dulwich into the `try:` block.

Reviewed By: ikostia

Differential Revision: D18225592

fbshipit-source-id: 207c797dee9cc439c84d3c997417d7545775075d