Fix encoding in repo converter
Summary: There are commits in art that contain that have non-ASCII bytes in their commit messages. We need to make sure to encode these gracefully because mononoke must have valid UTF-8 commit messages. We can use the converter_source.recode() function in the way that the git source does, but that doesn't seem to fix the full problem with import. This commit contains extra filtering in commit messages to scrub 0x7f-0xff bytes out.

Reviewed By: xavierd

Differential Revision: D18140370

fbshipit-source-id: 7b836efa91bd2fd99debec9e525d9b941023d224