mononoke/blobimport: do not process::exit(1)
Summary:
In general, this is a pretty bad practice because it can cause logs to go
missing since slog doesn't flush immediately.

More practically, I'm trying to figure out if this is related to the ASAN
failure jsgf has reported while running a test that causes blobimport to error
out (the LFS blobimport one, which is the only one that triggers a blobimport
error — on purpose).

Upon doing further investigation on this, it looks like this ASAN error
occassionaly triggers on blobimport errors (which I accidentally discovered in
a completely unrelated diff of mine that broke blobimport in CI:
https://our.intern.facebook.com/intern/testinfra/diagnostics/2814749782696535.562949955931708.1572443231/)

What's weird is that there is hardly anything happening in blobimport after
that point: we just `process::exit()` out.

So, my plan is to see if it continues showing up after this change, in order to
try and understand slightly better what's happening. To that end, I'd like to
land this diff, then check in a week if we had any failures on the test in
question.

Reviewed By: StanislavGlebik

Differential Revision: D18225593

fbshipit-source-id: 53be1b3bc0fcdbd33ac24575cacb0c7fd4aea5a7