mononoke: split off unbundle response generation
Summary:
Higher-level goal of this stack is to implement the "push redirector" -  a mechanism that takes a push to a specific (small in the megarepo terminology) repository, redirects it to run in a different (large) repository and waits until the result of this push it is backsynced into a small repository again. Note that "push" here is a generic term, meaning "unbundle wireproto command".

This particular diff is a necessary step of separating `unbunde` processing form generating `unbundle` response:
- processing will be done in the redirection target
- response generation needs to happen in the redirection source

To achieve it, the diff uses the same technique we used to separate `unbundle` parsing from processing:
- introduce an `enum` (`UnbundleResponse`), which encodes all possible results of processing (i.e. data, necessary to generate the response bytes)
- implement response generation for each variant of the `enum`
- this `enum` now can be converted between small- and large-repo pipelines

A byproduct of the changes I made in this diff is a bunch of changed `.context()` messaging. Hopefully, nothing is too controversial.

Reviewed By: krallin

Differential Revision: D18203805

fbshipit-source-id: 0cf22e63231a592333949b8b879d3e1aa855eb34