mononoke: extract marking pushrebased changesets as public from response gen
Summary:
Before, this piece of logic was erroneously located in the wireproto
response-generating functions. It worked because we would always generate
response for the pushrebased commits.

Today though, we plan to run pushrebase in one repo, do x-repo sync and
generate a response from the other repo. This means that marking as public
*has* to happen in the repo, where processing goes on.

Reviewed By: krallin

Differential Revision: D18244760

fbshipit-source-id: 203f3a502106a067c12bb2519d72735ae6c5d550