convert: fix test which uses gitutil.parsegitcommitraw
Summary:
D18140370 changed the signature of `gitutil.parsegitcommitraw`, which breaks
`test-fb-convert-repo.py` as it was using it.

The signature change meant the function now expects a source repo as an
argument.  This is a problem for the test, as it doesn't have one to hand.
It's also arguably overkill, as it adds a new dependency for this function,
where really it just wants a function to perform recoding with.

Change the function to expect a `recode` function (for which we can use
`source.recode`), so we can still allow the test to directly test the function.

Reviewed By: quark-zju

Differential Revision: D18247041

fbshipit-source-id: 4537d631af28862d183a604325732cc50af0829b