backingstore: use ContentStore to import tree
Summary:
This diff adds `get_tree` method to the backingstore crate to enable EdenFS to
import manifest data from hgcache.

This diff still uses the ctreemanifest code to parse the data we get from the
data store. This is going to be switched to use the Rust implementation in
upcoming diffs.

Reviewed By: xavierd

Differential Revision: D18097700

fbshipit-source-id: 271df49f31df1805e03406c60f6c24b9306f9434