types: alias HgId to Id20
Summary:
The `HgId` implements a lot of convenient features. Alias it to Id20 so we can
use it for other 20-byte stuff, like a non-hg SHA1 binary hash.

In the future we might want to make `HgId` and `Id20` incompatible types.
For example, define `struct HgId(Id20)`, and define deref, and handy constructors.
But that requires larger code change.

The name `Id20` is Stefan Filip's idea.

Reviewed By: sfilipco

Differential Revision: D18265059

fbshipit-source-id: 4f505c1d38d25b405f7588d76a7752c32c62f8be