Make hg rage provide useful directory/uptime/IP data on Windows
Summary:
On Windows, `hg rage` fails to provide useful information in these cases:
```
[1]: uptime error: 'uptime' is not recognized as an internal or external command,
operable program or batch file.
...
[2]: ifconfig error: 'ifconfig' is not recognized as an internal or external command,
operable program or batch file.
```
And unless msysgit/Cygwin is in the path, the 'ls' command fails too. This fixes all of the above to provide better rage support for Windows users.

Reviewed By: quark-zju

Differential Revision: D18271656

fbshipit-source-id: af2af7cc5346709b43bfd0dc87ab18f45dd6b672