remove the logic in fix-code.py for fixing copyright statements
Summary:
As of D18266092 we now apply Facebook's internal copyright header linters to
the Mercurial code.  This handles checking the copyright headers and providing
auto-fixed versions of the code.

The internal linter operates slightly differently than the logic in
`fix-code.py`: it handles separate copyrights for internal-only directories
and third-party directories, as well as multiple license headers in some
files.  Currently `fix-code.py` complains about the header required by the
internal linter for some files (particularly for internal-only files).  Rather
than trying to update `fix-code.py` to duplicate the logic of our internal
linters this removes the logic internally from `fix-code.py`.  This should
work well for Facebook internal-developers, although it means this check won't
automatically be visible for any external contributors.  For now this seems
like an okay trade-off for the simplicity.

Reviewed By: quark-zju

Differential Revision: D18280731

fbshipit-source-id: e153f272e8eaec24a1d497dc379405723a88238c