mononoke: extract hook running into an explicit step of the unbundle pipeline
Summary:
Before this diff, hook running was a part of unbundle processing, which means
that we hooks were run in the same repo, where the processing happened. This
is not desirable for the push-redirection scenarios, where we need to run hooks
in the small repo, which initially accepts an unbundle. This diff allows for
that possibility.

Note: I do realize how ugly the pipeline got. Once I have some time, I want to:
- move stuff from `client/unbundle/mod.rs` into `client/unbundle/processing.rs`
- move `bundle2_resovler` into `client/unbundle` nad maybe name it `parsing`?
- extract all pipeline-result struct definitions (`PostResolveAction`,
  `UnbundleResult` etc) into a separate file
- come up consistent naming for these structs

Reviewed By: StanislavGlebik

Differential Revision: D18270301

fbshipit-source-id: eb25302efefca54fa513a6a01149fae4d351886c