blackbox: make network entries with read_bytes & result readable
Summary:
Those two entries serialize to the same key (`R`), so if you have both,
deserialization fails and your entry is silenty gone (and if you had just a
result, I'm guessing it becomes `read_bytes`, which is probably not desirable
anyway).

This fixes this by serializing to 2 different keys.

Reviewed By: quark-zju

Differential Revision: D18272274

fbshipit-source-id: 0b09343430d0b8a9ef8325d8432b47cec6393337