edenapi: log Mononoke Session IDs to blackbox
Summary:
Earlier today, I had what seems like a very slow tree prefetch (took ~2 minutes for a single tree), but I wasn't really able to figure out what might have caused it (is a backend problem? something else?).

This diffs is a first step towards making this a little easier to debug: it updates edenapi to log Mononoke Session IDs (and a few other things) to the hg blackbox.

This also updates backingstore a little bit since bringing in the `regex` crate
seems to cause the Rust compiler to not be able to locate which type to use in
`assert_eq!` for a `vec![]` (presumably because there are multiple things that
can be compared to a `Vec<u8>`).

Reviewed By: kulshrax

Differential Revision: D17879330

fbshipit-source-id: a387a031fb3359ece3e3210bc09617290a1f9907