hgcommands: flush blackbox before returning from `run_command`
Summary:
Running `hg debughttp`, a lot of the output (e.g. command finish) ends up
missing from the blackbox, because we never flush it (normally, it flushes on
`Drop`, but when we call `process::exit()`, `Drop` implementations aren't
called).

This patch adds a flush before exiting from `run_command` (which seems like it
should be fairly cheap if the log is not dirty).

Reviewed By: quark-zju

Differential Revision: D18272273

fbshipit-source-id: b7b17f2a6e7c743f7db660a48640137276e890f0