pushrebase: expose real "old" commit to pushkey hooks if commits were not rewritten
Summary:
During the ovrsource write rollout today, we ran into an interesting race
condition in blobimport (which I'll document elsewhere in more detail), which
resulted in Mononoke pushing a commit to master that wasn't a descendant of
master (it lacked one commit, which we landed separately).

Prior to the push, master was at X here:

```
X <- master
|
A
```

Then, Mononoke attempted to push (this push is not a pushrebase), but its
understanding of master was incorrect, and it though A was master. This would
normally be blocked by our hooks to deny non-fast-forward pushed, but
unfortunately it turns out hat the push was in fact allowed.

This left us with a history looking like this:

```
C <- master
|
B X
|/
A
```

The root cause seems to be that the pushrebase code updates the pushkey data
that pushkey hooks end up looking at, even if a pushrebase wasn't performed.
This resulted in the hook preventing non-ff pushes allowing the push.

Reviewed By: quark-zju

Differential Revision: D18208700

fbshipit-source-id: 5b728b42a3324344114979f081454ebf74e276cc