mononoke: deleted files manifest data structures
Summary:
This is a first step in implementation of the Deleted files manifest (https://fb.quip.com/Wk4XAi7dyXRr).

Deleted Files Manifest is a data structure that tracks deleted files and commits where they were deleted. This manifest was designed in addition to Unodes to make following file history across deletions possible.

The diff introduces a core data structure for the manifest.

Reviewed By: StanislavGlebik

Differential Revision: D18204687

fbshipit-source-id: 3424127da68fb7818f4da5608e4115702a3c9092