Fix getchanges() correctness by returning new instead of old blobs
Summary:
In `getchanges()`, we are erroneously returning the old (i.e. parent tree's)
blob for each file changed by a commit. This behavior probably stemmed from
loose wording in the documentation for `getchanges()`, which calls for the
"source revision id" of each file. (That phrase doesn't really make sense, as
blobs aren't generally considered revisions and the terms "source" and
"destination" are generally only used for copies/moves.) However, it's clear
from what the function is supposed to do (report what changed in a revision)
that it has to return the new blob for each file or else there would be no
way for the caller to accurately reconstruct the commit.

Fix this to bring the behavior in line with other backends. This was a major
correctness issue.

Also fix the handling of moves so that we generate a deletion entry for the
old filename, again bringing us in line with other backends. And don't
generate identity entries in `copies` for files that have been modified but
not moved.

Reviewed By: mwdevine

Differential Revision: D18093896

fbshipit-source-id: 913f50d8b71bd0291806caf7ed47a4dea5718cf3