tracing-collector: use different EspanId for different TracingData
Summary:
Right now a TracingData uses 0-based EspanIds. There is no way to tell if a
EspanId belongs to a TracingData or another. This is not a problem if there is
only one TracingData. But I'd like to enable testing assertion use-cases like:

    tracer = tracingdata()
    with tracer:
        # log data to the specified tracer
        ...
    # make assertions about the data logged in tracer.
    # for example, "fetch_trees" were called "<how many>" times with
    # expected paths, and some paths are not fetched.

As an attempt to not mess up integrity of TracingData, let's allocate different
EspanId ranges for different TracingData.

Reviewed By: sfilipco

Differential Revision: D18130643

fbshipit-source-id: d21ace8ab099a9e95ec4353962826df043444073