pymanifest: improve matches performance
Summary:
`manifest.matches` returs a manifestdict. The manifestdict interface is used
in a few places in python code.
The implementation for manifestdict that we use is called by lazymanfiest and
is backed by an array. Insertions are O(N) when not done in an ascending order.
Rust Manifests wouldn't insert lines in order by default.
This change updates the constrution of the manfiestdict with a precomputed
flatmanifest.

Reviewed By: quark-zju

Differential Revision: D18312996

fbshipit-source-id: 04e16ae40bc9c7c8bdbc8c123629f8c2a1984b8b