revert notifing eden when a transaction is rolled back
Summary:
Summary.
This reverts D17943808, which updated the transaction abort code to use Eden's
`resetParentCommits()` API to point Eden at the old parent commits from before
the transaction started.

In practice this is bad in most cases.  If the working directory has
legitimately been updated to reflect a different commit's state, we should
leave the dirstate parent pointing there.  Attempting to restore the old
parent commit pointer leaves the dirstate commit pointing at something that
possibly extremely different from the current checkout contents.

This results in much worse behavior for users than the problem it was
originally trying to fix: with the new behavior this leaves user's checkouts
in a state where Mercurial commands take an extremely long time to run, and
may attempt to download the entire repository contents from the servers in
order to compute the status differences.

The problem that D17943808 was trying to fix is more rare, and the
consequences aren't as bad when it does occur.  D17943808 was trying to fix
the problem that sometimes the dirstate can point to a non-existent commit
after a transaction abort.  When this does occur, at least Mercurial is in a
state where commands fail immediately, and this is easily fixed by running
`eden doctor`.

It seems like the best way to fix this problem in the future is to make sure
that commits can never be deleted once they are created.  Aborting a
transaction should hide commits that were created by the transaction, but it
should not remove them from existence entirely.

Reviewed By: quark-zju

Differential Revision: D18320058

fbshipit-source-id: 656d252155ba7725cae42cd8c5b20fba71e39d0d