mononoke/lfs_server: bring a bit more sanity to Scuba logging
Summary:
We grew our Scuba logging sort of organically in the LFS Server, and it's starting to show. Notably, we have several places setting the same keys (on purpose, but that's brittle), and some inconsistency in terms of whether we're looking at a content length of an actual request or response size.

This diff fixes that by:

- Making sure we differentiate content lengths and response sizes,
- Moving all Scuba keys into an enum that has documentation,
- Adding an integration test for Scuba logging (by logging to a file),
- Adding sane defaults for various fields (content length, notably),
- For downloads, capturing the volume of data we actually sent.

This will cause a bit of churn in our Scuba tables, but this is probably the right time to do it. Notably, things that look at `response_size` and `upload_size` should now probably start looking at `response_bytes_sent` and `request_bytes_received` instead. I'll update our derived columns to look at those.

This also updates our ODS logging a little bit, but we don't have any alarms looking at content size yet.

Reviewed By: HarveyHunt

Differential Revision: D18303711

fbshipit-source-id: f7e955c872242dc49c379f24230a151eb9e25fda