mononoke: megarepo wc verifier
Summary:
A tool that checks that files in a commit from a large repo match files from a
corresponding commit in a small repo.

The tool does verification in a few stages:
1) Fetch paths and filenodes for small and large repo
2) Move all paths in the large repo
3) For each path check if we have the same path in small repo. If we don't
print a warning
4) For each filenode that are different also compare their content ids, and
repo an error only if content ids are different

Note that we can't rely just on filenode id comparison, because it turns out
that filenodes can legitimately be different between large and small repo.
We could also just compare the content ids, but that's much slower.

Reviewed By: krallin

Differential Revision: D18268939

fbshipit-source-id: 14f091db380e127ce2d926253376fbc0cf1e30b3