mononoke/lfs_server: add live configurations (/ configerator)
Summary:
This adds support in the Mononoke LFS Server for reloading configurations in
real time, and notably for tracking Configerator configurations.

Since we don't have async Rust bindings to Configerator (yet), this spawns a
thread that periodically checks Configerator for updates (which is probably
good enough).

This also adds support for hot-reloading configurations from files as well,
which I've used for integration tests here.

Reviewed By: HarveyHunt

Differential Revision: D18270434

fbshipit-source-id: 57527cd9f7388144022cb509fc2d695dc4e1c1df