hg grep: fixup biggrep integration to respect repo relative cwd
Summary:
This diff fixes up how the repo relative path is passed
down to the biggrep client.  Biggrep doesn't really 100% know about
paths and instead uses a regex to match filenames, so we're constructing
a regex alternation based on the list of filename matchers.

This isn't 100% perfect but it is usually good enough.

Reviewed By: simpkins

Differential Revision: D18318785

fbshipit-source-id: 845f4d186e768c45535ee3a611b6cbe2963329f2