mononoke: Consistent routing for LFS server downloads
Summary:
When clients request a blob from the LFS server they are randomly
assigned a server, which will fetch the blob from a backing store and then
place it into an in memory cache. Future requests for the blob will be serviced
from the cache and therefore very fast.

In order to improve the cache hit rate, we can send requests for the same blob
to the same host using consistent hashing. The consistent hashing is implemented
on the proxygen side, by looking for a URL parameter.

When clients make a batch request, append a URL parameter to the returned download
URLs that proxygen can use to consistently route the client to the same host.

Reviewed By: krallin

Differential Revision: D18298193

fbshipit-source-id: 07bbad1673df627a1be7f12fb06397d4f9732d36