hgcommands: remove some clones of arguments
Summary:
`OsString` implements `From<String>` but not `From<&str>`. But there is another
way to get the `OsString` - `OsString::push(AsRef<OsStr>)`.

The `push` overhead can be similar to clone a String. But the point is
`HgPython::new()` does not always need to create the `OsString`s. In chg's
case, the new code should be cheaper.

Reviewed By: sfilipco

Differential Revision: D18097877

fbshipit-source-id: 90debf33a700d314e1c7b021ea33182669e580d3