eden: improve detection of bind mounts in redirect fixup
Summary:
We've seen some weird situations where the VFS has some surprising
layering.  My favorite example so far had 4 layers of mounts, with the lower 3
layers being dead edenfs mounts that were still present in the mount table but
that had no associated fuse server process.

In that situation the mount table looked approximately like this:

```
/some/path/repo             <-- top of vfs
/some/path/repo/buck-out    invisible because of the mount above
/some/path/repo             an older instance
/some/path/repo/buck-out    invisible because of the mount above
/some/path/repo             an older instance
/some/path/repo/buck-out    invisible because of the mount above
/some/path/repo             the oldest instance
```

aside from being weird this shouldn't impact our behavior, except that it was
because our logic to uncover the effective bind mount configuration just looks
at the mount table and trusts that one of the matches mounts is visible to the
VFS.

The problem is that the top layer has no buck-out bind mount redirection which
leads to buck writing lots of data to the overlay.

To resolve this we need to verify that the mount table entry is visible
in the VFS and for that we can use our existing `is_bind_mount` helper
function.

Reviewed By: pkaush

Differential Revision: D18335468

fbshipit-source-id: 509c8b43f5e69f0a42c292a7011271d81ab97919